+++ erp/web/src/lang/en.ts
@@ -19,6 +19,10 @@ export default {
     supplierProduct: 'Supplier Product',
     supplierProductList: 'Supplier Product List',
     saveSupplierProduct: 'Create Supplier Product',
+    product: 'Product Setting',
+    productList: 'Product List',
+    newProduct: 'Create New Clietnt',
+    editProduct: 'Edit Client',
     documentation: 'Documentation',
     guide: 'Guide',
     permission: 'Permission',
