+++ erp/web/src/views/system/product/edit-product.vue
@@ -43,7 +43,7 @@ import { Component, Vue } from 'vue-property-decorator'
 import { ProductModule } from '@/store/modules/product'
 import '@/assets/custom-theme/index.css'
 import productName from '@/views/components/product-name.vue'
-import taxTypePulldown from '../components/tax-type-pulldown.vue'
+import taxTypePulldown from '@/views/components/tax-type-pulldown.vue'
 import money from '@/views/components/money.vue'
 import { Form as ElForm } from 'element-ui'
 
@@ -70,7 +70,7 @@ export default class extends Vue {
         await ProductModule.UpdateProduct(this.product)
         this.$router
           .push({
-            path: 'product'
+            path: 'product-list'
           })
           .catch(err => {
             console.warn(err)
